pyproject.toml
[project]
name = "pythonaiagenttechwithtim"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.17",
    "langchain-anthropic>=1.4.2",
    "langchain-core>=1.3.2",
    "langchain-openai>=1.2.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]
